TFG_IA_Signos-feature1/Backend/FastAPI/requirements.txt
fastapi==0.135.3
pydantic==2.13.0
supabase==2.28.3
fastapi==0.135.3
python-dotenv==1.2.2